fix(app): 31 Adjusted test to account for envar setting
This test has been failing for a while and I couldn't figure out why until now.  In manually testing the spinner spinner I set the AIA_VERBOSE envar to be true.  This resulted in AIA.config.verbose? to be true.  The test was originally expected to be false.  Now its expected value depends on the presents or absence of the AIA_VERBOSE envar.