Added the ruby-openai gem
The current crop of back-end processors do not have a speach capability.  Decided to do it within AIA using a direct connection to OpenAI's text to speech capability.

Updated the AIA.speak method to use either the default Mac command "say" when the AIA_VOICE is set to 'siri' and OS.osx? is true.

Added --voice and --speech_model to work with OpenAI's TTS endpoint.