AIA Architecture Sucks!
This programmed started life as a single file mind to keyboard script and then grew without a clear direction from there.  I need to throw everything out.  Update the PromptManager gem and the AiClient gem.  Once that is done I should be able to design a new version of AIA that is not so complicated to maintain.